[V2][Main UI] 기본 작업 완료. UI 작업 외에 누락됐던 월드 <-> GameInstance <-> 배틀 간의 PlayerData 복사도 같이 처리

+++ HOF/Source/HOF/Public/HOFPlayerState.h
@@ -43,7 +43,7 @@ public:
 	inline void ModifyFood(int32 amount) { PlayerData.Food(amount); }
 	inline void ModifyGold(int32 amount) { PlayerData.Gold(amount); }
 	inline void ModifyMaxHP(int32 amount) { PlayerData.HP.ModifyMaxValue(amount); }
-	inline void ModifyCurrentHP(int32 amount)
+	inline void ModifyCurrentHP(float amount)
 	{
 		PlayerData.HP(amount);
